+++ CHANGELOG.rst
@@ -1,3 +1,11 @@
+next version
+------------
+
+New features
+~~~~~~~~~~~~
+
+* Add modeling.calculate_se_from_cov
+
 0.48.0 (2022-02-04)
 -------------------